+++ R/write_plink.R
@@ -26,6 +26,7 @@
 #' This must be a length-\eqn{n} vector.
 #' This will be ignored (with a warning) if \code{fam} is provided.
 #' @param verbose If TRUE (default) function reports the paths of the files being written (after autocompleting the extensions).
+#' @param append If `TRUE`, appends loci onto the BED and BIM files (default `FALSE`).  In this mode, all individuals must be present in each write (only loci are appended); the FAM file is not overwritten if present, but is required at every write for internal validations.  If the FAM file already exists, it is not checked to agree with the FAM table provided.
 #'
 #' @return Invisibly, a named list with items in this order: X (genotype matrix), bim (tibble), fam (tibble).
 #' This is most useful when either BIM or FAM tables were auto-generated.
@@ -63,7 +64,7 @@
 #' \url{https://www.cog-genomics.org/plink/1.9/formats}
 #'
 #' @export
-write_plink <- function(file, X, bim = NULL, fam = NULL, pheno = NULL, verbose = TRUE) {
+write_plink <- function(file, X, bim = NULL, fam = NULL, pheno = NULL, verbose = TRUE, append = FALSE) {
     # die if things are missing
     if (missing(file))
         stop('Output file path is required!')
@@ -122,10 +123,11 @@ write_plink <- function(file, X, bim = NULL, fam = NULL, pheno = NULL, verbose =
 
     # now everything has been checked and/or created
     # time to write file!
-    write_bed(file, X, verbose = verbose)
-    write_fam(file, fam, verbose = verbose)
-    write_bim(file, bim, verbose = verbose)
-
+    write_bed(file, X, verbose = verbose, append = append)
+    write_bim(file, bim, verbose = verbose, append = append)
+    if ( !append || !file.exists( add_ext( file, 'fam') ) ) 
+        write_fam(file, fam, verbose = verbose)
+    
     # invisibly return the data passed and/or created
     return(
         invisible(