+++ NEWS.md
@@ -97,3 +97,7 @@
 # 2020-07-03 - genio 1.0.15.9000
 
 * Functions `write_plink`, `write_bed`, and `write_bim` now have `append` option, for writing extremely large files in parts.
+
+# 2020-07-28 - genio 1.0.16.9000
+
+* Added `write_eigenvec` to write plink/GCTA eigenvector files.